chore: add Enterprise Edge explanation
