+++ frontend/src/component/admin/enterprise-edge/EnterpriseEdgeInstances/EnterpriseEdgeInstance/EnterpriseEdgeInstance.tsx
