+++ frontend/src/component/admin/enterprise-edge/EnterpriseEdgeInstances/EnterpriseEdgeInstance/EnterpriseEdgeInstanceLatency.tsx
